Backend/create/search test
The basic structure of the backend is there and live on heroku. You can look at the routes and how i call them in the test app. (cobaltwebserver part)

Right now create saves and shows while search just shows all the trips in the database as testing and proof of concept stuff